Add files via upload
The library I created is an implementation of the REFPROP library for the C programming language. REFPROP is a widely-used library for performing thermodynamic calculations and fluid property evaluations. My implementation in C allows developers to directly utilize the functions and methods of REFPROP in their C programs.

The library offers a wide range of functions that span from basic property calculations such as pressure, temperature, enthalpy, and entropy, to more advanced calculations like determining thermodynamic regions, mixture properties, and specific substance properties.

With this library, developers can perform accurate and reliable calculations of thermodynamic properties of fluids using the equations and models provided by REFPROP. This enables them to design and optimize systems and processes involving fluids in a wide range of applications, including chemical engineering, refrigeration systems, energy, air conditioning, and more.

In summary, my library implements the functionality of REFPROP in the C language, providing developers with access to powerful calculation capabilities and thermodynamic properties for their applications.